backend/-generateService/requirements.txt
# Core FastAPI dependencies
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
PyJWT[crypto]>=2.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
prometheus-client>=0.19.0

# Image processing (optional - for future enhancements)
# Pillow>=10.0.0

# Utilities
pydantic>=2.7.0
httptools>=0.6.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0
